requirements.txt
flask
psutil
protobuf
requests
python-snappy
opentelemetry-api
opentelemetry-sdk
opentelemetry-proto
opentelemetry-exporter-prometheus-remote-write
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
